feat(basics-errors): recoverable and propagating errors
This demonstrates how to handle recoverable errors as well as propagating errors using
the Result<T,E> enum as well as using the ? operator.

Additionally, this shows how to use the unwrap(), expect() and the unwrap_or_else() methods
to handle recoverable errors as alternatives to use the match statement.