feat(algorithms, arrays): sorted squared array
Adds the algorithm for handling the sorted square array problem where an
input of integers is used as an input to the function and the output
result is the squares of each value in ascending order.

The assumption made here is that the input array is in ascending order
and that the integers can be negative, meaning that the order of the
squares will be the same as the input array elements that generated the
squares i.e. the positions might not match.

This uses a two pointer approach to ensure that the resultant array is
in order. The two pointer approach as the benefit of using less time to
handle sorting as sorting is not done in this approach. However a new
array is created which incurs a cost of O(n) where n is the size of the
input array.